backend/monolith/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.8

# Auth
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
pydantic[email]==2.7.1

# File processing
PyMuPDF==1.24.3

# AWS
boto3==1.34.110

# Embedding (sentence-transformers backend)
sentence-transformers==2.7.0
transformers==4.40.2
torch==2.3.0
numpy==1.26.4

# Vector store
faiss-cpu==1.8.0

# HTTP client (for Ollama)
httpx==0.27.0

# Google OAuth (optional — comment out if not needed)
google-auth==2.29.0
